+++ test-cases/test-spring-boot/src/main/java/com/hyf/hotrefresh/hello/invoke/TestTrigger.java
@@ -1,0 +1,7 @@
+package com.hyf.hotrefresh.hello.invoke;
+
+public class TestTrigger {
+    public static void trigger() {
+        TestInvoker.invoke();
+    }
+}